pyproject.toml
[project]
name = "llm-agent-template"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shota_kizawa", email = "shomok0106k@gmail.com" }
]
dependencies = [
    "langchain>=0.2.14",
    "openai>=1.42.0",
    "langchain-openai>=0.1.22",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/llm_agent_template"]
